+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Missing recipes for new building blocks
+
 ## [1.21.1-5.1.2] - 2025-07-13
 Requires Silent Gear 4.0.25 or higher
 ### Added